ChainModels.ChainModel — Type

ChainModel{T} = KChainModel{<:AbstractVector{<:AbstractArray{T,2}}}

A specialized KChainModel with K=2 nearest-neighbor interactions

ChainModels.FactorizedModel — Type

FactorizedModel{T} = KChainModel{<:AbstractVector{<:AbstractArray{T,1}}}

A specialized KChainModel with K=1 nearest-neighbor interactions, i.e. no interactions

ChainModels.KChainModel — Type

KChainModel{T,L} <: DiscreteMultivariateDistribution

A type to represent a discrete multivariate probability distribution factorized on a one-dimensional chain of length L.

FIELDS

f a vector containing the factors as arrays

ChainModels.accumulate_left! — Method

accumulate_left!(l, f::AbstractVector{<:AbstractArray{<:Real,K}}) where {K}

In-place version of accumulate_left

ChainModels.accumulate_left — Method

accumulate_left(f::Vector{Matrix{T}}) where {T<:Real}

Compute the left partial normalization for the matrices in f

\[\]

ChainModels.accumulate_middle! — Method

accumulate_middle!(l, f::Vector{Matrix{T}}) where {T<:Real}

In-place version of accumulate_middle

ChainModels.accumulate_middle — Method

accumulate_middle(f::Vector{Matrix{T}}) where {T<:Real}

Compute the middle partial normalization for the matrices in f

\[m_{i,j}(x_i,x_j) = \log\sum\limits_{x_{i+1},\ldots,x_{j-1}}\prod\limits_{k=i}^{j-1} e^{f_k(x_k,x_{k+1})}\]

ChainModels.accumulate_right! — Method

accumulate_right!(l, f::AbstractVector{<:AbstractArray{<:Real,K}}) where {K}

In-place version of accumulate_right

ChainModels.accumulate_right — Method

accumulate_right(f::Vector{Matrix{T}}) where {T<:Real}

Compute the right partial normalization for the matrices in f

\[\]

ChainModels.avg_energy — Method

avg_energy(chain::KChainModel; nmarg = neighbor_marginals(chain))

Compute the average energy

\[E = - \sum_{x_1, \ldots, x_L} \sum_{i=1}^{L-K+1} f_i(x_i,\ldots,x_{i+K-1}) p(x_1, \ldots, x_L)\]

ChainModels.compute_empirical_Kmarginals — Method

compute_empirical_Kmarginals(X::AbstractMatrix{<:Integer}, K::Integer; qs=tuple(maximum(X, dims=2)...), eps=0.0)

Compute K- and K-1-body empirical frequencies from data X which is a matrix LxN where L is the chain length and N the number of samples.

ChainModels.evaluate — Method

evaluate(chain::ChainModel, x)

Evaluate the (possibly unnormalized) model at x

ChainModels.lognormalization — Method

lognormalization(chain::KChainModel; l = accumulate_left(chain))

Conceptually equivalent to log(normalization(chain)), less prone to numerical issues

ChainModels.marginals — Method

marginals(chain::KChainModel; l = accumulate_left(chain), r = accumulate_right(chain))

Compute single-site marginals

\[p(x_i)\quad\forall i\]

Optionally pass pre-computed left and right partial normalizations.

ChainModels.nbody_neighbor_marginals — Method

nbody_neighbor_marginals(::Val{n}, chain::KChainModel; l = accumulate_left(chain), r = accumulate_right(chain))

Compute the n-body marginals

\[p(x_i, \ldots, x_{i+n-1})\quad\forall i\]

with $n\le K$. Optionally pass pre-computed left and right partial normalizations.

ChainModels.neighbor_marginals — Method

neighbor_marginals(chain::KChainModel; l = accumulate_left(chain), r = accumulate_right(chain))

Compute the K-body marginals

\[p(x_i, \ldots, x_{i+K-1})\quad\forall i\]

where K is the number of nearest-neighbors involved in the interactions. For a ChainModel, it computes the pairwise marginals $p(x_i, x_{i+1}$. Optionally pass pre-computed left and right partial normalizations.

ChainModels.normalization — Method

normalization(chain::KChainModel; l = accumulate_left(chain))

Compute the normalization of the model

ChainModels.nstates — Method

nstates(f::AbstractVector{<:AbstractArray{<:Real}})

Returns an iterator with the number of values each variable can take.

ChainModels.pair_marginals — Method

pair_marginals(chain::ChainModel{T};
    l = accumulate_left(chain), r = accumulate_right(chain), 
    m = accumulate_middle(chain)

Compute pairwise marginals

\[p(x_i, x_j)\quad\forall i,j\]

Optionally pass pre-computed left, right and middle partial normalizations.

ChainModels.rand_chain_model — Method

rand_chain_model(rng::AbstractRNG, L::Integer, q::Integer)

Return a ChainModel of length L and q states for each variable, with random entries

ChainModels.rand_factorized_model — Method

rand_factorized_model(rng::AbstractRNG, L::Integer, q::Integer)

Return a FactorizedModel of length L and q states for each variable, with random entries

ChainModels.rand_kchain_model — Method

rand_kchain_model([rng], K::Integer, L::Integer, q::Integer)

Return a KChainModel of length L and q states for each variable, with random entries

LinearAlgebra.normalize! — Method

LinearAlgebra.normalize!(chain::KChainModel; logZ = lognormalization(chain))

Normalize the distribution such that the normalization constant is equal to 1.