Types

ChainModels.ChainModel — Type

ChainModel{T,L} <: DiscreteMultivariateDistribution

A type to represent a discrete multivariate probability distribution factorized on a one-dimensional chain of length L.

FIELDS

f :: Vector{Matrix{T}} contains the factors

Functions

ChainModels.accumulate_left — Function

accumulate_left(f::Vector{Matrix{T}}) where {T<:Real}

Compute the left partial normalization for the matrices in f

\[l_{i}(x_{i+1}) = \sum\limits_{x_1,\ldots,x_i}\prod\limits_{j=1}^i e^{f_j(x_j,x_{j+1})}\]

ChainModels.accumulate_right — Function

accumulate_right(f::Vector{Matrix{T}}) where {T<:Real}

Compute the right partial normalization for the matrices in f

\[r_{i}(x_{i-1}) = \sum\limits_{x_i,\ldots,x_L}\prod\limits_{j=i-1}^L e^{f_j(x_j,x_{j+1})}\]

ChainModels.accumulate_middle — Function

accumulate_middle(f::Vector{Matrix{T}}) where {T<:Real}

Compute the middle partial normalization for the matrices in f

\[m_{i,j}(x_i,x_j) = \sum\limits_{x_{i+1},\ldots,x_{j-1}}\prod\limits_{k=i}^{j-1} e^{f_k(x_k,x_{k+1})}\]

ChainModels.accumulate_left! — Function

accumulate_left!(l, f::Vector{Matrix{T}}) where {T<:Real}

In-place version of accumulate_left

ChainModels.accumulate_right! — Function

accumulate_right!(l, f::Vector{Matrix{T}}) where {T<:Real}

In-place version of accumulate_right

ChainModels.accumulate_middle! — Function

accumulate_middle!(l, f::Vector{Matrix{T}}) where {T<:Real}

In-place version of accumulate_middle

ChainModels.nstates — Function

nstates(f::Vector{Matrix{T}}) where {T<:Real}

Returns an iterator with the number of values each variable can take.

nstates(chain::ChainModel{T,L})

Returns a NTuple{L,Int} with the number of values each variable can take.

Base.length — Function

Base.length(::ChainModel{T,L})

Returns L, the number of variables

ChainModels.evaluate — Function

evaluate(chain::ChainModel, x)

Evaluate the (possibly unnormalized) model at x

ChainModels.logevaluate — Function

logevaluate(chain::ChainModel, x)

Conceptually equivalent to log(evaluate(chain, x)), less prone to numerical issues

ChainModels.normalization — Function

normalization(chain::ChainModel; l = accumulate_left(chain))

Compute the normalization

\[Z = \sum\limits_{x_1,\ldots,x_L}\prod\limits_{i=1}^{L-1} e^{f_i(x_i,x_{i+1})}\]

Optionally, pass the pre-computed left partial normalization

ChainModels.lognormalization — Function

lognormalization(chain::ChainModel; l = accumulate_left(chain))

Conceptually equivalent to log(normalization(chain)), less prone to numerical issues

LinearAlgebra.normalize! — Function

normalize!(chain::ChainModel; logZ = lognormalization(chain))

Divides each factor by $Z^{1/L}$ so that the normalization becomes $1$.

LinearAlgebra.normalize — Function

normalize(chain::ChainModel; logZ = lognormalization(chain))

Return a new ChainModel equivalent to chain but rescaled so that normalization equal to $1$.

ChainModels.marginals — Function

marginals(chain::ChainModel; l = accumulate_left(chain), r = accumulate_right(chain))

Compute single-variable marginals $p(x_i=x)$

\[p(x_i=x) = \sum_{x_1, x_2, \ldots, x_L} p(x_1, x_2, \ldots, x_L) \delta(x_i,x)\]

Optionally, pass the pre-computed left and right partial normalization

ChainModels.neighbor_marginals — Function

neighbor_marginals(chain::ChainModel; l = accumulate_left(chain), r = accumulate_right(chain))

Compute nearest-neighbor marginals $p(x_i=x,x_{i+1}=x')$

\[p(x_i=x,x_{i+1}=x') = \sum_{x_1, x_2, \ldots, x_L} p(x_1, x_2, \ldots, x_L) \delta(x_i,x) \delta(x_{i+1},x')\]

Optionally, pass the pre-computed left and right partial normalization

ChainModels.pair_marginals — Function

neighbor_marginals(chain::ChainModel; l = accumulate_left(chain), r = accumulate_right(chain), m = accumulate_middle(chain))

Compute pair marginals $p(x_i=x,x_j=x')$

\[p(x_i=x,x_j=x') = \sum_{x_1, x_2, \ldots, x_L} p(x_1, x_2, \ldots, x_L) \delta(x_i,x) \delta(x_j,x')\]

Optionally, pass the pre-computed left, right and middle partial normalization

ChainModels.loglikelihood_gradient — Function

loglikelihood_gradient(chain::ChainModel, x; neigmarg = neighbor_marginals(chain))

Compute the gradient of the loglikelihood $\mathcal{L}(\boldsymbol{x})$ of samples $\{x^{(\mu)}\}_{\mu=1,\ldots,M}$ with respect to the functions chain.f

\[\frac{d\mathcal{L}(\boldsymbol{x})}{d f_i(y_i,y_{i+1})} = \sum_{\mu=1}^M \delta(y_i,x_i^{(\mu)})\delta(y_{i+1},x_{i+1}^{(\mu)}) - M p(y_i,y_{i+1}) \]

Optionally pass pre-computed neighbor marginals

ChainModels.loglikelihood_gradient! — Function

loglikelihood_gradient!(df, chain::ChainModel, x; neigmarg = neighbor_marginals(chain)

In-place version of loglikelihood_gradient