ChainModels.ChainModel — Type

ChainModel{T,L} <: DiscreteMultivariateDistribution

A type to represent a discrete multivariate probability distribution factorized on a one-dimensional chain of length L.

FIELDS

f :: Vector{Matrix{T}} contains the factors

Base.length — Method

Base.length(::ChainModel{T,L})

Returns L, the number of variables

ChainModels.accumulate_left! — Method

accumulate_left!(l, f::Vector{Matrix{T}}) where {T<:Real}

In-place version of accumulate_left

ChainModels.accumulate_left — Method

accumulate_left(f::Vector{Matrix{T}}) where {T<:Real}

Compute the left partial normalization for the matrices in f

\[l_{i}(x_{i+1}) = \log\sum\limits_{x_1,\ldots,x_i}\prod\limits_{j=1}^i e^{f_j(x_j,x_{j+1})}\]

ChainModels.accumulate_middle! — Method

accumulate_middle!(l, f::Vector{Matrix{T}}) where {T<:Real}

In-place version of accumulate_middle

ChainModels.accumulate_middle — Method

accumulate_middle(f::Vector{Matrix{T}}) where {T<:Real}

Compute the middle partial normalization for the matrices in f

\[m_{i,j}(x_i,x_j) = \log\sum\limits_{x_{i+1},\ldots,x_{j-1}}\prod\limits_{k=i}^{j-1} e^{f_k(x_k,x_{k+1})}\]

ChainModels.accumulate_right! — Method

accumulate_right!(l, f::Vector{Matrix{T}}) where {T<:Real}

In-place version of accumulate_right

ChainModels.accumulate_right — Method

accumulate_right(f::Vector{Matrix{T}}) where {T<:Real}

Compute the right partial normalization for the matrices in f

\[r_{i}(x_{i-1}) = \log\sum\limits_{x_i,\ldots,x_L}\prod\limits_{j=i-1}^L e^{f_j(x_j,x_{j+1})}\]

ChainModels.energy — Method

energy(chain::ChainModel; nmarg = neighbor_marginals(chain))

Compute the "energy"

\[\begin{aligned} E =& \mathbb{E} \sum_{x_1, x_2, \ldots, x_L} \left[ -\sum\limits_{i=1}^{L-1}f_i(x_i,x_{i+1}) \right] p(x_1, x_2, \ldots, x_L)\\ =& -\sum\limits_{i=1}^{L-1} \sum_{x,x'} f_i(x,x') p(x_i=x,x_{i+1}=x') \end{aligned}\]

Optionally, pass the pre-computed neighbor marginals

ChainModels.evaluate — Method

evaluate(chain::ChainModel, x)

Evaluate the (possibly unnormalized) model at x

ChainModels.logevaluate — Method

logevaluate(chain::ChainModel, x)

Conceptually equivalent to log(evaluate(chain, x)), less prone to numerical issues

ChainModels.lognormalization — Method

lognormalization(chain::ChainModel; l = accumulate_left(chain))

Conceptually equivalent to log(normalization(chain)), less prone to numerical issues

ChainModels.marginals — Method

marginals(chain::ChainModel; l = accumulate_left(chain), r = accumulate_right(chain))

Compute single-variable marginals

\[\begin{aligned} p(x_i=x) &= \sum_{x_1, x_2, \ldots, x_L} p(x_1, x_2, \ldots, x_L) \delta(x_i,x)\\ =& \frac1Z e^{l_{i-1}(x)+r_{i+1}(x)} \end{aligned}\]

Optionally, pass the pre-computed left and right partial normalization

ChainModels.neighbor_marginals — Method

neighbor_marginals(chain::ChainModel; l = accumulate_left(chain), r = accumulate_right(chain))

Compute nearest-neighbor marginals

\[\begin{aligned} p(x_i=x,x_{i+1}=x') =& \sum_{x_1, x_2, \ldots, x_L} p(x_1, x_2, \ldots, x_L) \delta(x_i,x) \delta(x_{i+1},x')\\ =& \frac1Z e^{l_{i-1}(x) + f_i(x,x') + r_{i+2}(x')} \end{aligned}\]

Optionally, pass the pre-computed left and right partial normalization

ChainModels.normalization — Method

normalization(chain::ChainModel; l = accumulate_left(chain))

Compute the normalization

\[\begin{aligned} Z =& \sum\limits_{x_1,\ldots,x_L}\prod\limits_{i=1}^{L-1} e^{f_i(x_i,x_{i+1})}\\ =& \sum\limits_{x_L} e^{l_{L-1}(x_L)} \end{aligned}\]

Optionally, pass the pre-computed left partial normalization

ChainModels.nstates — Method

nstates(f::Vector{Matrix{T}}) where {T<:Real}

Returns an iterator with the number of values each variable can take.

ChainModels.nstates — Method

nstates(chain::ChainModel{T,L})

Returns a NTuple{L,Int} with the number of values each variable can take.

ChainModels.pair_marginals — Method

pair_marginals(chain::ChainModel; l = accumulate_left(chain), r = accumulate_right(chain), m = accumulate_middle(chain))

Compute pair marginals

\[\begin{aligned} p(x_i=x,x_j=x') =& \sum_{x_1, x_2, \ldots, x_L} p(x_1, x_2, \ldots, x_L) \delta(x_i,x) \delta(x_j,x')\\ =& \frac1Z e^{l_{i-1}(x) + m_{i,j}(x,x') + r_{j+1}(x')} \end{aligned}\]

Optionally, pass the pre-computed left, right and middle partial normalization

LinearAlgebra.normalize! — Method

normalize!(chain::ChainModel; logZ = lognormalization(chain))

Divide each factor by $Z^{1/L}$ so that the normalization becomes $1$.

LinearAlgebra.normalize — Method

normalize(chain::ChainModel; logZ = lognormalization(chain))

Return a new ChainModel equivalent to chain but rescaled so that normalization equal to $1$.